BookSmith/requirements.txt
Package              Version
-------------------- -------
asgiref              3.2.10
Django               3.1.2
django-widget-tweaks 1.4.8
Pillow               8.0.1
pip                  20.2.4
pytz                 2020.1
setuptools           50.3.1
sqlparse             0.4.1
wheel                0.35.1
